--- a/bin/.gitignore
+++ b/bin/.gitignore
@@ -1,1 +1,0 @@
-/SudokuSolverGUI.class