--- a/bin/.gitignore
+++ b/bin/.gitignore
@@ -1,0 +1,1 @@
+/test.class