Added comments, helper methods and improved cohesion.