@brandon1024 2017-02-09
Performance Improvements:
To help decrease the number of iterations needed to solve a given puzzle, at the time of constructing the SudokuGrid object, each cell is assigned an array of possible cell values. This is computed based on the other initial cell values. This helps to decrease the number of invalid cell values tested at runtime.

If a given cell has only one single possible cell value, the cell is assigned that value and marked as unmodifiable. Subsequently, the search for possible cell values is restarted to update the possible cell values, and also look for other cells with only one possible cell value. This step can significantly decrease the difficulty of the puzzle, and simplify the search for a solution to the puzzle.

The new algorithm has been proven more efficient than the previous. Below are the benchmark results:
```
Test Puzzle:
| 8 ¦ 0 ¦ 0 | 0 ¦ 0 ¦ 0 | 0 ¦ 0 ¦ 0 |
| 0 ¦ 0 ¦ 3 | 6 ¦ 0 ¦ 0 | 0 ¦ 0 ¦ 0 |
| 0 ¦ 7 ¦ 0 | 0 ¦ 9 ¦ 0 | 2 ¦ 0 ¦ 0 |
| 0 ¦ 5 ¦ 0 | 0 ¦ 0 ¦ 7 | 0 ¦ 0 ¦ 0 |
| 0 ¦ 0 ¦ 0 | 0 ¦ 4 ¦ 5 | 7 ¦ 0 ¦ 0 |
| 0 ¦ 0 ¦ 0 | 1 ¦ 0 ¦ 0 | 0 ¦ 3 ¦ 0 |
| 0 ¦ 0 ¦ 1 | 0 ¦ 0 ¦ 0 | 0 ¦ 6 ¦ 8 |
| 0 ¦ 0 ¦ 8 | 5 ¦ 0 ¦ 0 | 0 ¦ 1 ¦ 0 |
| 0 ¦ 9 ¦ 0 | 0 ¦ 0 ¦ 0 | 4 ¦ 0 ¦ 0 |

Previous Algorithm Results:
Time Elapsed: 72ms
Outer Loop Iterations: 182491
Inner Loop Iterations: 484158

New Algorithm Results:
Time Elapsed: 65ms
Outer Loop Iterations: 144111
Inner Loop Iterations: 328110

Outer Loop Iterations Improvement: 21%
Inner Loop Iterations Improvement: 32%
```